Fix admin panel 500: bigint subscriber count fails integer validation
listAllSources() (backing GET /api/admin) COUNT()s subscribers per
source. Postgres COUNT() returns bigint, which the JS driver
serializes as a string to avoid precision loss beyond
Number.MAX_SAFE_INTEGER -- COALESCE(sc.count, 0) doesn't change that.
The TypeBox response schema expects subscriberCount as an integer, so
parseSourceListRows() threw "Database returned invalid source list
rows" for every row, on every request, unconditionally. Reproduced via
a real admin account (the default dev-DB user, and the one from an
existing browser test fixture, weren't admins, which is presumably why
this had gone unnoticed) and confirmed the fix end-to-end.

Cast to ::int in the query -- subscriber counts never approach
int4's ~2.1 billion ceiling.

+++ src/db/data-services/source-data-service.ts
@@ -210,7 +210,7 @@ export class SourcesDataService {
         WITH subscriber_counts AS (
             SELECT
                 us.source_id,
-                COUNT(us.user_id) AS count
+                COUNT(us.user_id)::int AS count
             FROM user_sources AS us
             GROUP BY us.source_id
         )
